Added Edit mode for editing the visibilty of the profile icons
- Disabled unhandled exception catching when running under VS
- fixes #192

To edit, go to settings and click "Edit Profiles List". Once in this mode clicking on a profile icon will cause it to become grayed out. This will result in it being hidden from view once you leave the edit mode by clicking the same button. To show it again do the same thing except it will become fully visible indicating that it will be shown when the edit mode is left.